requirements.txt
cassandra-driver==3.29.0
colbert-ai==0.2.19
Flask==3.0.2
langchain-text-splitters==0.0.1
sentence-transformers==2.5.1
torch==2.2.1